pyproject.toml
[project]
name = "mla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cuda-tile>=1.2.0",
    "cupy-cuda13x>=14.0.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
docs = [
    "myst-parser>=5.0.0",
    "sphinx>=9.1.0",
    "sphinx-rtd-theme>=3.1.0",
]

[tool.uv]
default-groups = ["docs"]
